Complete Task 2

+++ src/com/oop/Lab1/Task2/Course.java
@@ -1,16 +1,18 @@
 package com.oop.Lab1.Task2;
+
 import java.util.*;
+
 public class Course {
     Scanner in = new Scanner(System.in);
     public String c_name;
-    public int mark;
+    public double mark;
 
-    public Course()
-    {
+    public Course() {
         System.out.print("    Course name: ");
-        c_name=in.nextLine();
+        c_name = in.nextLine();
         System.out.print("    Mark: ");
-        mark=in.nextInt();
+        mark = in.nextDouble();
         System.out.println();
     }
+
 }
